Move STET tests to their own module
